move suspense and error boundary to widget chrome
