--- routes/dashboard/widget-dashboard/components/widget-render/widget-render.module.css
@@ -1,10 +1,0 @@
-.loading,
-.error {
-	height: 100%;
-	padding: var(--wpds-dimension-padding-md);
-	color: var(--wpds-color-fg-content-neutral-weak);
-}
-
-.error {
-	text-align: center;
-}
