+++ routes/dashboard/widget-dashboard/components/widget-render/widget-render.tsx
@@ -1,66 +1,14 @@
 /**
- * External dependencies
- */
-import type { ReactNode } from 'react';
-
-/**
  * WordPress dependencies
  */
-import { Component, Suspense, useCallback } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
-import { Stack } from '@wordpress/ui';
+import { useCallback } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import { useDashboardInternalContext } from '../../context/dashboard-context';
 import { getLazyWidgetComponent } from '../../utils/get-lazy-widget-component';
-import styles from './widget-render.module.css';
 import type { DashboardWidget, WidgetType } from '../../types';
-
-interface ErrorBoundaryProps {
-	children: ReactNode;
-}
-
-interface ErrorBoundaryState {
-	hasError: boolean;
-}
-
-class WidgetErrorBoundary extends Component<
-	ErrorBoundaryProps,
-	ErrorBoundaryState
-> {
-	state: ErrorBoundaryState = { hasError: false };
-
-	static getDerivedStateFromError(): ErrorBoundaryState {
-		return { hasError: true };
-	}
-
-	render() {
-		if ( this.state.hasError ) {
-			return (
-				<Stack
-					direction="column"
-					justify="center"
-					align="center"
-					className={ styles.error }
-					role="alert"
-				>
-					<p>{ __( 'This widget encountered an error.' ) }</p>
-				</Stack>
-			);
-		}
-		return this.props.children;
-	}
-}
-
-function LoadingOverlay() {
-	return (
-		<Stack justify="center" align="center" className={ styles.loading }>
-			<span>{ __( 'Loading…' ) }</span>
-		</Stack>
-	);
-}
 
 interface WidgetRenderInternalProps {
 	widget: DashboardWidget< unknown >;
@@ -96,24 +44,23 @@ function WidgetRenderImpl( { widget, widgetType }: WidgetRenderInternalProps ) {
 	);
 
 	return (
-		<WidgetErrorBoundary>
-			<Suspense fallback={ <LoadingOverlay /> }>
-				{ /* WidgetComponent is a cached `lazy()` keyed by renderModule, so its identity stays stable across renders. */ }
-				{ /* eslint-disable-next-line react-hooks/static-components */ }
-				<WidgetComponent
-					attributes={ widget.attributes }
-					setAttributes={ setAttributes }
-				/>
-			</Suspense>
-		</WidgetErrorBoundary>
+		<>
+			{ /* WidgetComponent is a cached `lazy()` keyed by renderModule, so its identity stays stable across renders. */ }
+			{ /* eslint-disable-next-line react-hooks/static-components */ }
+			<WidgetComponent
+				attributes={ widget.attributes }
+				setAttributes={ setAttributes }
+			/>
+		</>
 	);
 }
 
 /**
- * Lazy-loads a widget's render module via the configured resolver and renders
+ * Resolves a widget's render module via the configured resolver and renders
  * it with the minimal `WidgetRenderProps` contract: `attributes` plus
- * `setAttributes`. Wraps the module in a `Suspense` boundary and an error
- * boundary so neighbours stay mounted if one widget fails.
+ * `setAttributes`. Suspense and error handling live one layer up in the
+ * `WidgetDashboard.Widget` chrome, so a failing or pending body does not tear
+ * down the surrounding header or controls.
  *
  * Kept internal to the package. Surfaces that want bare widget rendering
  * should compose `WidgetDashboard.Widget` instead.
