+++ routes/dashboard/widget-dashboard/components/widget/widget.module.css
@@ -2,3 +2,14 @@
 	height: 100%;
 	box-sizing: border-box;
 }
+
+.loading,
+.error {
+	height: 100%;
+	padding: var(--wpds-dimension-padding-md);
+	color: var(--wpds-color-fg-content-neutral-weak);
+}
+
+.error {
+	text-align: center;
+}
